Adjust a few scancodes in the translation table
Tab is more useful than Delete, so map the Aspect Ratio key to the
former. Instead of using the same Play/Pause code for both the Play
and Pause keys, map the latter to Space, which is a useful key to
have and just so happens to also function as the Play/Pause toggle
in most media players. Special internet/multimedia keys make for
better custom shortcuts than plain function keys as they are more
distinctive even without modifiers, so map the four color keys to
some of those instead.

Also, remove some symbolic constants for scancodes which Linux does
not recognize out of the box, without loading a custom scancode map.
There are more than enough directly supported scancodes available to
cover all buttons on a remote control, and desktop shortcuts can be
used to reassign key functions as needed.