Allow keycodes to be combined with modifiers
Widen scancode table entries to 14 bit, i.e. full program words.
This allows for various flag bits to be combined with the base
scancode. Use four flag bits to support any combination of the
modifier keys Shift, Control, Alt and Super to be sent together
with the key's base scancode.