Stop waiting for argument on new keyboard command
If we receive a new command from the PS/2 host while expecting
an argument byte to a previous command, abort the last command
and handle the new one. This should help to avoid getting into
a stuck state expecting an argument that never arrives.
To determine whether a received argument byte is actually a new
command, make use of the fact that any valid argument must have
a value below E0. As a special case, the resend command will be
executed without clearing the argument expected state.