Implement basic key auto-repeat
Some environments, such as Windows with default settings, the Linux text
console, or various boot manager interfaces still expect the keyboard to
implement typematic auto-repeat in hardware. Therefore, for the sake of
out-of-the-box compatibility, implement rudimentary key auto-repeat in
the PS/2 IR adapter.
In order to keep things simple, the repeat rate is fixed and tied to the
repeat rate used by the IR remote control. The delay until auto-repeat
begins can be changed by the host via the set rate/delay command of the
keyboard protocol. The timing is not particularly accurate, but good
enough.
More precise as well as more flexible control of the typematic behavior
can be obtained by activating the software auto-repeat implementations
of modern desktop environments.